requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-genai==1.12.1
python-dotenv==1.0.1
slowapi==0.1.9
